+++ apps/server/src/api/sustainability/commands/update-settings.command.ts
@@ -1,0 +1,10 @@
+import { SustainabilityModel } from "../models/sustainability.model"
+
+export default async function updateSustainabilitySettingsCommand(userId: string,
+  useDarkMode: boolean,
+  useFastestNode: boolean,
+  useLessEnergy: boolean,
+  useOptimizedAPICalls: boolean) {
+  await SustainabilityModel.findOneAndUpdate({ userId }, { useDarkMode, useFastestNode, useLessEnergy, useOptimizedAPICalls })
+  return true
+}